requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and analysis
pandas==2.0.3
numpy==1.24.3

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.21.0

# Technical analysis
ta==0.10.2

# Data validation
pydantic==2.5.0

# Optional: For enhanced performance
# ta-lib==0.4.28  # Uncomment if you have TA-Lib installed
